requirements.txt
langchain>=1.0.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pytest>=8.0.0
